feat: implement MCP tool utility functions and refactor Anthropic and OpenAI providers to utilize them
# Conflicts:
#	src/renderer/src/providers/OpenAIProvider.ts
